pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["controllable_patching_striding"]

[project]
name = "controllable_patching_striding"
requires-python = ">=3.8"
version = "0.0.1"
readme = "README.md"
dependencies = [
    "torch>=2.1,<3",
    "h5py>=3.9.0,<4",
    "hydra-core>=1.3,<2",
    "einops>=0.8,<0.9",
    "wandb>=0.17.9,<0.18",
    "torch_harmonics>=0.6,<0.7",
    "neuraloperator==0.3.0",
    "timm>=1.0,<2",
    "matplotlib>=3.8,<4",
    "plotly>=5.0,<6",
    "numpy>=2.0,<3",
    "imageio>=2.25,<3",
    "torchinfo>=1.8.0,<2",
    "pandas>=2.2,<3",
    "the_well[benchmark] @ git+https://github.com/PolymathicAI/the_well.git@master",
]

[project.scripts]
the-well-download = "the_well.utils.download_script:main"

[project.optional-dependencies]
test = [
    "pytest",
    "ruff",
    "isort[colors]>=5.13.2",
    "pytest-order>=1.3.0",
]

[tool.isort]
profile = "black"

[tool.mypy]
ignore_missing_imports = true
